refactor: make non-instantiable parent classes abstract
