+++ ors-api/src/main/java/org/heigit/ors/api/responses/routing/IndividualRouteResponse.java
@@ -19,7 +19,7 @@ import org.heigit.ors.api.requests.routing.RouteRequest;
 import org.heigit.ors.routing.RouteResult;
 import org.locationtech.jts.geom.Coordinate;
 
-public class IndividualRouteResponse {
+public abstract class IndividualRouteResponse {
     protected Coordinate[] routeCoordinates;
     protected boolean includeElevation = false;
     protected boolean isPtRequest = false;
